PS/2 Protocol - transmission working proof of concept in synchronous code
Plain bit banging code, did help to detect a hardware problem with the initial development setup.
The setup had series resistor in the DA and CK lines, but the internal pull-up on CK of the keyboard was stronger so the code could not issue a request to send correctly.

Current series resistor values: 330 Ohm.
Assumed keyboard pull up on CK: 2K Ohm.

Can be as low as 1K Ohm as per the references. The DA line had a much weaker pull up so it works fine. The keyboard supports USB so I assume the different values are for indicating Low speed / High speed mode support to the controller during USB initialization.

Files:
"PS2_test_sync_write.ino"